Remove superfluous generators, add competitive benchmarks
Now that I've done the testing to confirm Pcg64Mcg is the measurably higher perf for the cases I bothered to test I am removing the other algorithms to reduce library bloat and pave the way for a next pass that will remove the now unnecessary RngType enum.

Also took the time to add basic benchmarks of the other rust ksuid library which makes some different choices layout wise. As expected this library sorts faster and my effort in finding a high perf rng keeps creation on par, the other library goes to and from base62 faster.